backend/requirements.txt
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Cors==3.0.10
Flask-MySQL==1.5.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
PyMySQL==1.0.2
python-dotenv==0.21.0
six==1.16.0
Werkzeug==2.0.2
